Move stuff into chain subcommand
